Refactor ActiveSearch and BulkUpdate endpoints to unify endpoint interface naming and improve endpoint mapping consistency
